requirements.txt
dash==2.0.0
Werkzeug==2.0.0
dash-bootstrap-components==1.0.1
dash-bootstrap-templates==1.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
pandas==1.3.5
plotly==5.3.1
pyzenhub==0.2.1
PyGithub==1.55
spacy==3.2.4 # run: $ python3 -m spacy download pt_core_news_sm 
jupyterlab==3.3.4
matplotlib==3.5.2
seaborn==0.11.2
streamlit==1.3.1
click==7.1.2
github3.py==3.2.0
unidecode==1.3.4
kaleido==0.2.1
